Address cloud review: saveWin teardown UB, bench hardening
- Save-preset modal callback now goes through a SafePointer and the
  Content destructor dismisses the dialog first - the AlertWindow
  destructor fires the modal callback synchronously, which touched
  members mid-teardown if the host closed the editor with the dialog up.
- Update-check connection timeout 4s -> 2s (bounds the destructor join).
- BoRBench links juce_recommended_warning_flags like the plugin target;
  fixed the sign-conversion warnings that surfaced in FuzzChain.h.
- bor-bench setParam fails loudly on an unknown parameter id instead of
  jassert (a no-op in release) followed by a null deref - a stale id
  would have silently mis-calibrated the fuzz trims.

Review finding #3 (BoRBench missing JUCE_USE_CURL=0/JUCE_WEB_BROWSER=0)
was already satisfied - both defines are on the target.

+++ Source/FuzzChain.h
@@ -37,7 +37,7 @@ struct FuzzChain
         os->initProcessing ((size_t) blockSize);
 
         juce::dsp::ProcessSpec mono { sampleRate, (juce::uint32) blockSize, 1 };
-        for (int b = 0; b < 8; ++b)
+        for (size_t b = 0; b < 8; ++b)
         {
             gritBP[b].prepare (mono);
             gritBP[b].coefficients = juce::dsp::IIR::Coefficients<float>::makeBandPass (sr, CENTRES[b], 1.4f);
@@ -92,8 +92,8 @@ struct FuzzChain
             const float d2 = clip (d, 3.0f, asym, hard);       // grittier cascade
             const float diff = d2 - d;                          // extra harmonics to sprinkle per band
             float grit = 0.0f;
-            for (int b = 0; b < 8; ++b)
-                grit += g[(size_t) b] * gritBP[b].processSample (diff);
+            for (size_t b = 0; b < 8; ++b)
+                grit += g[b] * gritBP[b].processSample (diff);
             x[i] = (d + grit) * env[(size_t) i];                // restore the note's dynamics
         }
     }
